Upgrade to webpack 5
