+++ .eslintrc.yml
@@ -1,129 +1,4 @@
-parser: babel-eslint
-plugins:
-  - jest
-env:
-  browser: true
-  node: true
-  es6: true
-  jest/globals: true
-settings:
-  react:
-    version: '16.0'
 extends:
-  - 'eslint:recommended'
-  - 'prettier'
-  - 'plugin:react/recommended'
-  - 'plugin:jest/recommended'
-parserOptions:
-  ecmaVersion: 7
-  ecmaFeatures:
-    jsx: true
-  sourceType: module
+  - '@redhat-cloud-services/eslint-config-redhat-cloud-services'
 globals:
-  beforeAll: true
-  describe: true
-  expect: true
-  global: true
-  insights: true
-  it: true
-  mount: true
-  process: true
-  render: true
-  shallow: true
-  React: true
-rules:
-  array-bracket-spacing:
-    - error
-    - always
-    - objectsInArrays: false
-  comma-dangle: 2
-  comma-spacing:
-    - 2
-    - after: true
-  comma-style: 2
-  camelcase: 2
-  curly:
-    - error
-    - all
-  dot-notation: 2
-  eol-last: 2
-  eqeqeq: 2
-  func-names:
-    - error
-    - never
-  indent:
-    - error
-    - 4
-    - SwitchCase: 1
-      MemberExpression: 0
-      ImportDeclaration: 1
-      ObjectExpression: 1
-  react/jsx-curly-spacing:
-    - error
-    - always
-    - children: true
-    - allowMultiline: false
-    - spacing:
-      - objectLiterals:
-        - never
-  key-spacing: 2
-  keyword-spacing: 2
-  linebreak-style:
-    - error
-    - unix
-  max-len:
-    - 2
-    - 150
-  new-cap: 2
-  no-bitwise: 2
-  no-else-return: 2
-  no-caller: 2
-  no-mixed-spaces-and-tabs: 2
-  no-multiple-empty-lines:
-    - error
-    - max: 1
-  no-trailing-spaces: 2
-  no-use-before-define:
-    - error
-    - functions: false
-  no-undef: 2
-  no-unused-vars: 2
-  no-var: 2
-  no-with: 2
-  object-shorthand: 2
-  object-curly-spacing:
-    - error
-    - always
-    - objectsInObjects: false
-      arraysInObjects: false
-  one-var:
-    - error
-    - never
-  padding-line-between-statements:
-    - error
-    - blankLine: always
-      prev: block-like
-      next: "*"
-  prefer-const: 2
-  quote-props:
-    - error
-    - as-needed
-  quotes:
-    - error
-    - single
-    - allowTemplateLiterals: true
-  semi:
-    - error
-    - always
-  space-before-blocks: 2
-  space-in-parens: 2
-  space-infix-ops: 2
-  space-unary-ops:
-    - error
-    - words: false
-      nonwords: false
-  vars-on-top: 2
-  wrap-iife: 2
-  yoda:
-    - error
-    - never
+  insights: readable
