+++ config/dev.webpack.config.js
@@ -6,6 +6,14 @@ const { config: webpackConfig, plugins } = config({
     https: true
 });
 
+const modulesConfig = require('@redhat-cloud-services/frontend-components-config/federated-modules')(
+    {
+        root: resolve(__dirname, '../')
+    }
+);
+
+plugins.push(modulesConfig);
+
 module.exports = {
     ...webpackConfig,
     plugins
