+++ config/prod.webpack.config.js
@@ -4,6 +4,14 @@ const { config: webpackConfig, plugins } = config({
     rootFolder: resolve(__dirname, '../')
 });
 
+const modulesConfig = require('@redhat-cloud-services/frontend-components-config/federated-modules')(
+    {
+        root: resolve(__dirname, '../')
+    }
+);
+
+plugins.push(modulesConfig);
+
 module.exports = {
     ...webpackConfig,
     plugins
