+++ package.json
@@ -18,9 +18,7 @@
     "@redhat-cloud-services/sources-client": "^1.0.68",
     "awesome-debounce-promise": "2.1.0",
     "axios": "^0.18.1",
-    "babel-polyfill": "^6.26.0",
     "classnames": "^2.2.5",
-    "create-react-class": "^15.6.3",
     "react-hooks-testing-library": "^0.3.8",
     "react-syntax-highlighter": "12.2.1",
     "urijs": "^1.19.2"
@@ -58,40 +56,26 @@
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-flow": "^7.10.1",
     "@babel/preset-react": "^7.10.1",
-    "@redhat-cloud-services/frontend-components-config": "^2.1.6",
+    "@redhat-cloud-services/eslint-config-redhat-cloud-services": "0.0.1",
+    "@redhat-cloud-services/frontend-components-config": "3.0.0-beta12",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
-    "babel-loader": "^7.1.5",
+    "babel-loader": "^8.2.2",
     "babel-plugin-dual-import": "^1.2.1",
     "babel-plugin-lodash": "^3.3.2",
     "babel-plugin-transform-imports": "^2.0.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "clean-webpack-plugin": "^0.1.19",
-    "connect-history-api-fallback": "^1.6.0",
-    "copy-webpack-plugin": "^4.5.1",
-    "cross-fetch": "^2.2.2",
-    "css-loader": "^0.28.11",
-    "d3": "^5.16.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "enzyme-to-json": "^3.5.0",
-    "eslint": "^5.16.0",
+    "eslint": "^7.15.0",
     "eslint-config-prettier": "^2.10.0",
-    "eslint-loader": "^2.2.1",
-    "eslint-plugin-jest": "^22.21.0",
-    "eslint-plugin-react": "^7.20.0",
-    "file-loader": "^1.1.11",
-    "git-revision-webpack-plugin": "^3.0.6",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-jest": "^24.1.3",
+    "eslint-plugin-react": "^7.21.5",
     "glob": "^7.1.6",
-    "html-replace-webpack-plugin": "^2.5.6",
-    "html-webpack-plugin": "^3.0.4",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^23.5.0",
-    "koa-connect": "^2.0.1",
-    "lodash-webpack-plugin": "^0.11.5",
-    "material-ui": "^0.20.2",
-    "mini-css-extract-plugin": "^0.4.5",
-    "moment": "^2.26.0",
     "node-sass": "^4.14.1",
     "npm-run-all": "^4.1.5",
     "react": "^16.13.1",
@@ -101,19 +85,13 @@
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-promise-middleware": "^5.1.1",
-    "sass-loader": "^7.3.1",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^0.21.0",
     "stylelint": "^9.10.1",
     "stylelint-config-recommended-scss": "^3.3.0",
     "stylelint-scss": "^3.17.2",
-    "topojson": "^3.0.2",
-    "webpack": "^4.43.0",
+    "webpack": "^5.10.0",
     "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^3.3.11",
-    "webpack-serve": "^1.0.2",
-    "webpack-visualizer-plugin": "^0.1.11",
-    "write-file-webpack-plugin": "^4.5.1"
+    "webpack-cli": "^4.2.0",
+    "webpack-dev-server": "^3.11.0"
   },
   "scripts": {
     "build": "webpack --config config/prod.webpack.config.js",
@@ -124,7 +102,7 @@
     "lint:js": "eslint config src",
     "lint:sass": "stylelint 'src/**/*.scss' --config .stylelintrc.json",
     "server:ctr": "node src/server/generateServerKey.js",
-    "start": "NODE_ENV=development webpack-serve --config config/dev.webpack.config.js",
+    "start": "NODE_ENV=development webpack serve --config config/dev.webpack.config.js",
     "travis:build": "NODE_ENV=production webpack --config config/test.webpack.config.js",
     "travis:verify": "npm-run-all build:prod lint test",
     "verify": "npm-run-all build lint test",
