+++ src/bootstrap-dev.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import logger from 'redux-logger';
+import Remediations from './AppEntry';
+
+ReactDOM.render(<Remediations logger={ logger } />, document.getElementById('root'));
