+++ src/bootstrap.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Remediations from './AppEntry';
+
+ReactDOM.render(<Remediations />, document.getElementById('root'));
